Added StatementResult#summary to obtain the summary in the end of a result stream by pulling all unconsumed records into memory
