+++ driver/src/main/java/org/neo4j/driver/internal/InternalStatementResult.java
@@ -297,6 +297,17 @@ public class InternalStatementResult implements StatementResult
     }
 
     @Override
+    public ResultSummary summary()
+    {
+        while( !done )
+        {
+            connection.receiveOne();
+        }
+
+        return summary;
+    }
+
+    @Override
     public void remove()
     {
         throw new ClientException( "Removing records from a result is not supported." );
